[walnut] move LoadDPL/Workflow to load.go
